Clean unused serialization code
